+++ Model/PhormixChain.php
@@ -1,0 +1,135 @@
+<?php
+
+namespace Phormix\Model;
+
+use MVC\DataType\DTFileUpload;
+use MVC\DataType\DTRequestIn;
+use MVC\DataType\DTRoute;
+use Phormix\DataType\DTPhormixChain;
+use Phormix\DataType\DTPhormixSetup;
+
+class PhormixChain
+{
+    /**
+     * @var DTPhormixChain
+     */
+    protected $oDTPhormixChain;
+
+    /**
+     * @return void
+     */
+    protected function setStep()
+    {
+        (false === isset($_SESSION['ChainStep']))
+            ? $_SESSION['ChainStep'] = 0
+            : false
+        ;
+    }
+
+    /**
+     * @param \MVC\DataType\DTRequestIn $oDTRequestIn
+     * @return void
+     * @throws \ReflectionException
+     */
+    protected function jumpForms(DTRequestIn $oDTRequestIn)
+    {
+        // jump between forms by query param "step"
+        (false === empty($oDTRequestIn->get_queryArray()['step'] ?? ''))
+            ? $_SESSION['ChainStep'] = (int) ($oDTRequestIn->get_queryArray()['step'] - 1)
+            : false
+        ;
+    }
+
+    #-------------------------------------------------------------------------------------------------------------------
+    # public
+
+    /**
+     * @param \MVC\DataType\DTRequestIn        $oDTRequestIn
+     * @param \Phormix\DataType\DTPhormixChain $oDTPhormixChain
+     * @throws \ReflectionException
+     */
+    public function __construct(DTRequestIn $oDTRequestIn, DTPhormixChain $oDTPhormixChain)
+    {
+        $this->oDTPhormixChain = $oDTPhormixChain;
+        $this->setStep();
+        $this->jumpForms($oDTRequestIn);
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getStep()
+    {
+        $this->setStep();
+        return $_SESSION['ChainStep'];
+    }
+
+    /**
+     * @return \Phormix\Model\Phormix
+     * @throws \ReflectionException
+     */
+    public function getPhormix()
+    {
+        $oPhormix = Phormix::init(
+            DTPhormixSetup::create()
+                ->set_sConfigYamlFile($this->oDTPhormixChain->get_aDTPhormixSetup()[$_SESSION['ChainStep']]->get_sConfigYamlFile())
+                ->set_sElementDirectory($this->oDTPhormixChain->get_aDTPhormixSetup()[$_SESSION['ChainStep']]->get_sElementDirectory())
+                ->set_sValidateClass($this->oDTPhormixChain->get_aDTPhormixSetup()[$_SESSION['ChainStep']]->get_sValidateClass())
+        );
+
+        $oPhormix->run(bResetOnEmpty: false);
+
+        return $oPhormix;
+    }
+
+    /**
+     * @param \MVC\DataType\DTRoute  $oDTRoute
+     * @param \Phormix\Model\Phormix $oPhormix
+     * @return \Phormix\Model\Phormix
+     * @throws \ReflectionException
+     */
+    public function setActionOnRoutePath(DTRoute $oDTRoute, Phormix $oPhormix)
+    {
+        // make sure action ist just the route path (no query param "step")
+        $oPhormix->aConfig['form']['action'] = $oDTRoute->get_path();
+
+        return $oPhormix;
+    }
+
+    /**
+     * @param \Phormix\Model\Phormix $oPhormix
+     * @return \Phormix\Model\Phormix
+     * @throws \ReflectionException
+     */
+    public function proceed(Phormix $oPhormix)
+    {
+        // Form was successfully sent; Validation succeeded
+        if (true === $oPhormix->bSuccess)
+        {
+            // save
+            $_SESSION['Chain'][$_SESSION['ChainStep']]['sFormIdentifier'] = $oPhormix->sFormIdentifier;
+            $_SESSION['Chain'][$_SESSION['ChainStep']]['aData'] = $oPhormix->getDataAccepted();
+            (false === empty($_FILES)) ? $_SESSION['Chain'][$_SESSION['ChainStep']]['aFiles'] = $_FILES : false;
+
+            // call next formular
+            if ($_SESSION['ChainStep'] < (count($this->oDTPhormixChain->get_aDTPhormixSetup()) - 1))
+            {
+                $_SESSION['ChainStep']++;
+                $oPhormix = $this->getPhormix();
+            }
+        }
+
+        return $oPhormix;
+    }
+
+    /**
+     * @param \Phormix\Model\Phormix $oPhormix
+     * @return void
+     */
+    public function reset(Phormix $oPhormix)
+    {
+        unset($_SESSION['ChainStep']);
+        unset($_SESSION['Chain']);
+        $oPhormix->reset(bForce: true);
+    }
+}