+++ etc/config/Phormix/config/_function.php
@@ -4,11 +4,13 @@ if (false === function_exists('tidyMarkup'))
 {
     function tidyMarkup(string $sMarkup = '')
     {
-        $oDomHTMLDocument = \DOM\HTMLDocument::createFromString($sMarkup, LIBXML_HTML_NOIMPLIED);
-        $sMarkup = $oDomHTMLDocument->saveHTML();
+//        $oDomHTMLDocument = \DOM\HTMLDocument::createFromString($sMarkup, LIBXML_HTML_NOIMPLIED);
+//        $sMarkup = $oDomHTMLDocument->saveHTML();
 
         // remove multiple whitespaces
         $sMarkup = preg_replace('!\s+!', ' ', trim($sMarkup));
+        $sMarkup = str_replace('< ', '<', $sMarkup);
+        $sMarkup = str_replace(' >', '>', $sMarkup);
 
         return $sMarkup;
     }